create Docker container that runs the simulator
This commit adds a Dockerfile that builds a docker image for
running the DASHIF live source simulator. It is based upon the
Alpine Linux distribution. It uses Apache2 and mod_wsgi to
serve the manifest and fragments. The container exposes its
service on port 80. You can map that to any port that you
want. In the following example, port 8123.

   docker build -t dash-live .

   docker run -p 8123:80 -t dash-live